PR review - refactored incoming outgoing flows layout
