+++ activiti-cloud-modeling-service/activiti-cloud-services-modeling/activiti-cloud-services-modeling-service/src/main/java/org/activiti/cloud/services/modeling/validation/process/BpmnModelIncomingOutgoingFlowValidator.java
@@ -16,12 +16,9 @@
 package org.activiti.cloud.services.modeling.validation.process;
 
 import org.activiti.bpmn.model.BpmnModel;
-import org.activiti.bpmn.model.EndEvent;
 import org.activiti.bpmn.model.FlowNode;
-import org.activiti.bpmn.model.StartEvent;
 import org.activiti.cloud.modeling.api.ModelValidationError;
 import org.activiti.cloud.modeling.api.ValidationContext;
-import org.apache.commons.collections.CollectionUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -32,15 +29,11 @@ import java.util.stream.Stream;
  */
 public class BpmnModelIncomingOutgoingFlowValidator implements BpmnModelValidator{
 
-    public static final String NO_INCOMING_FLOW_PROBLEM = "Intermediate Flow node has no incoming flow";
-    public static final String NO_INCOMING_FLOW_PROBLEM_DESCRIPTION = "Intermediate Flow node has to have an incoming flow";
-    public static final String NO_OUTGOING_FLOW_PROBLEM = "Intermediate Flow node has no outgoing flow";
-    public static final String NO_OUTGOING_FLOW_PROBLEM_DESCRIPTION = "Intermediate Flow node has to have an outgoing flow";
-    public static final String INCOMING_FLOW_ON_START_EVENT_PROBLEM = "Intermediate Flow node Start event should not have incoming flow";
-    public static final String INCOMING_FLOW_ON_START_EVENT_PROBLEM_DESCRIPTION = "Intermediate Flow node Start event has to have an empty incoming flow";
-    public static final String OUTGOING_FLOW_ON_END_EVENT_PROBLEM = "Intermediate Flow node End event should not have outgoing flow";
-    public static final String OUTGOING_FLOW_ON_END_EVENT_PROBLEM_DESCRIPTION = "Intermediate Flow node End event should not have outgoing flow";
-    public static final String INTERMEDIATE_FLOW_VALIDATOR_NAME = "BPMN Intermediate Flow node validator";
+    private final List<FlowNodeFlowsValidator> flowNodeFlowsValidators;
+
+    public BpmnModelIncomingOutgoingFlowValidator(List<FlowNodeFlowsValidator> flowNodeFlowsValidators) {
+        this.flowNodeFlowsValidators = flowNodeFlowsValidators;
+    }
 
     @Override
     public Stream<ModelValidationError> validate(BpmnModel bpmnModel, ValidationContext validationContext) {
@@ -56,41 +49,10 @@ public class BpmnModelIncomingOutgoingFlowValidator implements BpmnModelValidato
     private List<ModelValidationError> validateTaskFlow(FlowNode flowNode) {
 
         List<ModelValidationError> errors = new ArrayList<>();
-        if (flowNode instanceof StartEvent) {
-            if (CollectionUtils.isEmpty(flowNode.getOutgoingFlows())) {
-                errors.add(createModelValidationError(NO_OUTGOING_FLOW_PROBLEM,
-                    NO_OUTGOING_FLOW_PROBLEM_DESCRIPTION,
-                    INTERMEDIATE_FLOW_VALIDATOR_NAME));
-            }
-            if (CollectionUtils.isNotEmpty(flowNode.getIncomingFlows())) {
-                errors.add(createModelValidationError(INCOMING_FLOW_ON_START_EVENT_PROBLEM,
-                    INCOMING_FLOW_ON_START_EVENT_PROBLEM_DESCRIPTION,
-                    INTERMEDIATE_FLOW_VALIDATOR_NAME));
-            }
 
-        } else if (flowNode instanceof EndEvent) {
-            if (CollectionUtils.isEmpty(flowNode.getIncomingFlows())) {
-                errors.add(createModelValidationError(NO_INCOMING_FLOW_PROBLEM,
-                    NO_INCOMING_FLOW_PROBLEM_DESCRIPTION,
-                    INTERMEDIATE_FLOW_VALIDATOR_NAME));
-            }
-            if (CollectionUtils.isNotEmpty(flowNode.getOutgoingFlows())) {
-                errors.add(createModelValidationError(OUTGOING_FLOW_ON_END_EVENT_PROBLEM,
-                    OUTGOING_FLOW_ON_END_EVENT_PROBLEM_DESCRIPTION,
-                    INTERMEDIATE_FLOW_VALIDATOR_NAME));
-            }
-
-        } else {
-            if (CollectionUtils.isEmpty(flowNode.getIncomingFlows())) {
-                errors.add(createModelValidationError(NO_INCOMING_FLOW_PROBLEM,
-                    NO_INCOMING_FLOW_PROBLEM_DESCRIPTION,
-                    INTERMEDIATE_FLOW_VALIDATOR_NAME));
-            }
-
-            if (CollectionUtils.isEmpty(flowNode.getOutgoingFlows())) {
-                errors.add(createModelValidationError(NO_OUTGOING_FLOW_PROBLEM,
-                    NO_OUTGOING_FLOW_PROBLEM_DESCRIPTION,
-                    INTERMEDIATE_FLOW_VALIDATOR_NAME));
+        for (FlowNodeFlowsValidator validator: flowNodeFlowsValidators) {
+            if (validator.canValidate(flowNode)){
+                errors.addAll(validator.validate(flowNode, this));
             }
         }
         return errors;
